Bump stripe ver

+++ phoenix-scala/phoenix/app/phoenix/models/payment/creditcard/CreditCardCharge.scala
@@ -66,7 +66,7 @@ object CreditCardCharge {
                      chargeId = stripe.getId,
                      state = Auth,
                      currency = currency,
-                     amount = stripe.getAmount)
+                     amount = stripe.getAmount.toInt) // FIXME int type shrink here
 }
 
 class CreditCardCharges(tag: Tag) extends FoxTable[CreditCardCharge](tag, "credit_card_charges") {
